+++ src/phosphorus/__main__.py
@@ -1,5 +1,6 @@
 from phosphorus.lib.parser import parse_args
 from phosphorus.subcommands.build import BuildCommand
+from phosphorus.subcommands.install import InstallCommand
 from phosphorus.subcommands.lock import LockCommand
 
 
@@ -8,5 +9,7 @@ def main() -> None:
     match args.subcommand:
         case "build":
             BuildCommand(args).run()
+        case "install":
+            InstallCommand(args).run()
         case "lock":
             LockCommand(args).run()