+++ src/phosphorus/lib/types.py
@@ -5,6 +5,7 @@ from typing import TYPE_CHECKING, TypedDict
 if TYPE_CHECKING:
     from phosphorus.lib.packages import Package
     from phosphorus.lib.requirements import ResolvedRequirement
+    from phosphorus.lib.resolver import LockEntry
 
 Match = str | None
 
@@ -13,3 +14,8 @@ class ResolvedPackageInfo(TypedDict):
     requirement: ResolvedRequirement
     hashes: tuple[str, ...]
     parents: set[Package]
+
+
+class PackageDiff(TypedDict):
+    update: set[tuple[ResolvedRequirement | None, LockEntry]]
+    remove: set[ResolvedRequirement]