+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 -   Added a subcommand to build sdists and wheels
 -   Added a subcommand to lock the dependencies
+-   Added a subcommand to install the project
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html